+++ Guardfile
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-guard :rspec, cmd: 'bundle exec rspec' do
+guard :rspec, cmd: 'bundle exec rspec --format documentation' do
   watch(%r{^spec/(.+)\.rb$})  { 'spec' }
   watch(%r{^lib/(.+)\.e?rb$}) { 'spec' }
 end
+
