--- lib/use_case/fetch_board.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-class FetchBoard
-  attr_reader :state
-
-  def initialize(board_gateway)
-    @board_gateway = board_gateway
-    @state = []
-  end
-
-  def execute
-    @board_gateway.fetch_board
-  end
-end
