+++ spec/test_doubles/board_gateway_stub.rb
@@ -1,0 +1,7 @@
+class BoardGatewayStub
+  attr_writer :board
+
+  def fetch_board
+    @board
+  end
+end
