+++ spec/unit/evaluate_board_spec.rb
@@ -1,48 +1,48 @@
 # frozen_string_literal: true
 
 describe EvaluateBoard do
-  class BoardGatewayStub
-    attr_writer :board
-
-    def fetch_board
-      @board
-    end
-  end
 
   let(:board_gateway) { BoardGatewayStub.new }
   let(:evaluate_board) { EvaluateBoard.new(board_gateway) }
 
   it 'continues the game until winning combination or all squares filled' do
+    board_gateway.board = Board.new(size: 9)
     expect(evaluate_board.execute).to eq(:continue)
   end
 
-  context 'has the winner' do
-    it 'when player one marks three in a horizontal row' do
+  context 'when there is a winner' do
+    it 'evaluates winning marks in a horizontal row' do
       board_gateway.board = ['X', 'X', 'X', nil, nil, nil, nil, 'O', 'O']
 
       expect(evaluate_board.execute).to eq(:player_one_wins)
     end
 
-    it 'when player one marks three in a vertical row' do
+    it 'evaluates winning marks in a vertical row' do
       board_gateway.board = ['X', nil, nil, 'X', nil, nil, 'X', 'O', 'O']
 
       expect(evaluate_board.execute).to eq(:player_one_wins)
     end
 
-    it 'when player one marks three in a vertical row (example 2)' do
+    it 'evaluates winning marks in a vertical row (example 2)' do
       board_gateway.board = [nil, 'X', nil, 'O', 'X', 'O', nil, 'X', nil]
 
       expect(evaluate_board.execute).to eq(:player_one_wins)
     end
 
-    it 'when player two marks three in a horizontal row' do
+    it 'evaluates winning marks in a horizontal row for other player' do
       board_gateway.board = ['O', 'O', 'O', nil, nil, nil, nil, 'X', 'X']
 
       expect(evaluate_board.execute).to eq(:player_two_wins)
     end
 
-    it 'when a player marks three in a diagonal' do
+    it 'evaluates winning marks in a diagonal' do
       board_gateway.board = [nil, nil, 'O', nil, 'O', nil, 'O', 'X', 'X']
+
+      expect(evaluate_board.execute).to eq(:player_two_wins)
+    end
+
+    it 'evaluates winning marks in a negative diagonal' do
+      board_gateway.board = ['O', nil, nil, nil, 'O', nil, 'X', 'X', 'O']
 
       expect(evaluate_board.execute).to eq(:player_two_wins)
     end
